Update dependencies in pnpm-lock.yaml file
Updated several packages in the pnpm lock file. This includes upgrading the lockfile version, TypeScript, Prisma, and other packages' versions. Removed several dependencies marked as deprecated or unrequired, such as '@highlight-run/next' and '@highlight-run/react'. There are also changes in the workspace dependencies. The main purpose of these changes is to leverage the latest capabilities of these libraries, improve the project's security and performance, and eliminate potential issues with outdated packages.