Remove redundant eslint configurations and refactor some imports
This commit streamlines the ESLint configuration by removing redundant files and updating related dependencies and scripts. The .eslintrc.js files in the root and apps/docs directories are removed as they are not necessary. Also, the 'lint' script is removed from the package.json file since we are not using it. Furthermore, import statements in admission/validator.ts and auth/layout.tsx are reorganized for better coding practices. At last, the details about your installed packages and their dependencies are updated in pnpm-lock.yaml.