Introduce 'use client' line in Toast.tsx
This commit adds 'use client' at the top of Toast.tsx file to enable client environment specification in the toast component. This improves the file's clarity and respects the ESLint rule structure. The 'use client' directive ensures that the script is run in a client-side context, lending clarity and specificity to its role and functionality.