Remove organization related API/module and Prettier config changes
Deleted the organization related components such as the creation of institutions as they are no longer needed for the project's use-case.
Also, upgraded Prisma from version 5.2.0 to 5.3.1 to leverage new features and improvements.
The Prettier configuration was also modified to consolidate formatting settings, moving all plugin dependencies to the end of the dependencies list, to improve consistency and readability of codebase.