Remove enrolment page from web app
The enrolment page was deleted to improve modularity and streamline the application structure. This page was managing form fetches and user sessions directly. By eliminating it, we have simplified the codebase, and maintenance will be less prone to mistakes. The services previously managed by this page are now correctly reassigned within the app framework. This contributes to a more coherent application architecture and improved efficiency in future development.