Add static branchId to admission form fetch
Added a static branchId to the getFormByCriteria function in the form page of the web app. This is a temporary change until the function can be updated to fetch the branchId from the user's session or profile. This will allow the form page to correctly fetch the relevant admission forms for the specific branch the user is associated with.