Update React versions in pnpm-lock file
Lowered the version of '@types/react' from 18.2.22 to 18.2.21 in several dependencies in the pnpm lock file.
Some deprecated or unnecessary dependencies were also removed, as they were no longer in use.

This version downgrade and cleaning was needed to ensure compatibility and avoid conflicts with other parts of the code base running on the downgraded version.