Refactor domain models and update branch service logic
Migrated existing models used in context with branch actions from the api folder directly into the domain folder. This move is a preparatory step towards implementing domain-driven design. In addition, the implementation of updating branches has been improved. It was extended to not only update the branch name and activation status, but also to record timestamps of creation and updates and connect the action with the user performing it. This provides more accurate and detailed data as well as more opportunities for data tracking and analysis.