Sharesheet Flyover API integration. (#184)
* Refactor Admission Share Flyout and related functionalities

Updated the UI of the Admission Share Flyout and moved it to '(protected)/admission/add/components' directory to better organize code structure. Refactored helper function to fetch data related to admission form share details and adjusted the service and route files accordingly. Furthermore, removed the unused 'ShareFlyout' component, and related 'getShareById' function in 'share/[id]/service.ts'. Updated the default styling of Button in Button.tsx to enhance consistency across the application. Updated AdmissionForm.tsx to replace the old sharing functionality with the new AdmissionShareFlyout component.

* Update AdmissionForm and AdmissionShareFlyout components

Include shareId in AdmissionFormProps and modify AdmissionShareFlyout to handle additional functionality, such as setting acceptPayment, linkExpiresOnDate, actualAmount, discountAmount, and isLinkActive state. This change also implements conditional rendering on shareId, thus the share details are only fetched if shareId is not null. Renamed date state to activeFromDate for better context in AdmissionShareFlyout.

The changes are supposed to provide more functionality to the user and make the code cleaner and easier to understand.

* Add features and refactor in Admission and Share modules

Modified the AdmissionForm to remove shareId, which was unused, and replaced the functionality to fetch share details of a form inside the component instead of passing it as a prop. The AdmissionShareFlyout component was also updated to fetch the share details and save them on click, making it more functional. "linkExpiresOnDate" state was renamed to "activeToDate", aligning it with the CreateShareModal type. An endpoint was also added to save the share.

This refactoring ensures the component is more self-contained and improves modularity. It also provides a clearer pathway for fetching and saving the share.

* Remove unused import in AdmissionPage.tsx

The import statement for 'getShareByFormId' service is removed as it's no longer used in the AdmissionPage component. This change simplifies the component by making it focus solely on admission logic and not dealing with share information directly. In order to retrieve the share details of a form, it's recommended to do so within a Share related component such as AdmissionShareFlyout.