Refine form share feature and error handling (#195)
Implemented the ability to create and update share details for a form. Refactored both schema and service for activeFromDate and activeToDate to accept string type instead of date, allowing the conversion to be performed within the service. This change improved data type consistency throughout the app. Fixed error status code in API/share/route.ts from 404 to 400 where applicable.

In the app, added more comprehensive instructions on the AdmissionShareFlyout. Handled possible mutation errors, providing an error toast. Provided a better approach for displaying and changing the active dates with a properly formatted date showing on the date labels. Also, implemented a copy-to-clipboard functionality for the share URL. Commit also includes renaming of some functions and endpoints, as well as adjustments in other related files for compatibility.