Refactor code and enhance security measures
This commit refactors some of our React components and enhances application security. We've streamlined the import paths for the Footer component in OtpVerifyForm.tsx and RecoveryPasswordForm.tsx. Additionally, we have made our course and department list components more semantic and accessible, wrapping the table headers in a 'thead' element.

We've also renamed functions to be more descriptive of their actions - 'DeleteCourseHandler' is now 'handleOnDeleteCourseClick'.

These changes make the code base more readable and maintainable. On the security front, we have added authorization checks to our API endpoints, reducing the risk of unauthorized access. This provides a more robust way to handle security threats.