[WIP] Schema changes for Student/Section/Test/Class/Batch/Branch/Organization (#198)
* Schema changes for Student/Section/Test/Class/Batch/Branch/Organization

To manage student details across different sections, classes, batches, branches and organization, added new models with necessary relations

* Remove students page and update sidebar data

The deletion of the students page was made due to its redundancy and an update was made to the 'sidebar/data.ts' file, altering the item IDs and paths to reflect this change. For instance, 'admission-addnew1' is now changed to 'students-addnew', and the corresponding path has been updated. This ensures consistency in our code base and improves navigation within the application.

* Refactor ID for 'Add New' option in sidebar data

Renamed the id from 'students-addnew' to 'students-add-new' in sidebar data for better readability and consistency in ID naming convention. This does not affect how the sidebar functions, but improves code maintainability.

* Update MongoDB credentials in .env file

The MongoDB credentials used in .env file were previously outdated/incorrect. Updated with the new password and database server URL to correct connection issues and ensure proper database interaction.

* Refactor student data model and related components

Optimized and restructured the student data model for more efficient data handling. Deleted the model file model.ts file from the api/student directory as it was redundant and moved its content to the lib/domain directory ensuring central access to all components that require it.
Read-only attribute added to all fields in the data.ts file in the api/onboarding/account directory and their maximum length increased to allow longer entries.
Prisma schema was modified to reflect changes in the data model. Now the fields accept null entries and some non-mandatory fields were deleted. The addStudent function now also creates a new user and assigns them to the proper organization and branch.
The route.ts file was also updated to check for an authenticated session before allowing a student to be added.
The changes were motivated by the need for a cleaner, more efficient data handling procedure for student data.

* students-list-ui completed (#203)

Co-authored-by: gopikumarpgk <gopikukmarcs443@gmail.com>

* Update dependency next-intl to v3.0.3 (#204)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* Update dependency next-intl to v3.1.0 (#206)

* Update dependency eslint to v8.54.0 (#207)

* student-list integration (#205)

* student-list integration

* Qodana error check

* studet list menu added

* Qodana warning cleared

---------

Co-authored-by: gopikumarpgk <gopikukmarcs443@gmail.com>

* Refactor code for readonly properties

* Refactor admission route and service

* Refactor form submission in AdmissionForm and
AddStudentForm

* Refactor user creation and add conditional check
for existing user

* Refactor getCourseList function parameters

* Remove organizationId parameter from getCourseList
function

* Refactor API service functions

* Update department and regulation service to remove
organizationId parameter

* Fixed build issue

* Remove unnecessary parameter in getRegulationList call

The organizationId parameter has been removed from the getRegulationList call in the regulation/page.tsx component. This is due to a recent rework of the student data model for better performance and streamlined related components where the organizationId became unnecessary, eliminating redundancies. The rework also includes changes to the Pr

---------

Co-authored-by: arunselvakumar <b.arunselvakumar@gmail.com>
Co-authored-by: gopikumarpgk <115463448+gopikumarpgk@users.noreply.github.com>
Co-authored-by: gopikumarpgk <gopikukmarcs443@gmail.com>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>