Add nextjs-toploader and exported primary color from UI package (#273)
* Add nextjs-toploader and exported primary color from UI package

This commit introduces the 'nextjs-toploader' to apps/web by updating the dependencies in 'package.json' and 'pnpm-lock.yaml', which is used to show a loading bar at the top of the screen. It also exports a constant, 'primaryColor', in 'packages/ui/colors.ts' for usage across the application, provided that it is correctly imported and utilized in 'apps/web/app/(protected)/layout.tsx'. The primaryColor is also used as the color for the NextTopLoader.

* Add NextTopLoader component to layout