Don't review (#281)
* Add regulation flyout bug fixing completed

* review comments addressed

* review comments addressed

* review comments addressed

* Style fixes

* Update validation logic in SignUpForm component

This commit removes the 'validatePhoneNumber' function and instead inline the validation logic directly in the phone number field settings inside the 'useForm' hook in the SignUpForm.tsx component. The 'validate' property of the phone number field now only allows for values that are exactly 10 digits long. Additionally, the form validation mode is changed to ‘onBlur’ to improve user experience.

* margin added

---------

Co-authored-by: Arun Selva Kumar <b.arunselvakumar@gmail.com>