Delete Confirmation Modal  (#312)
* ...Switching Branch

* ...Delete Confirmation Button Design Done

* ...Run Format Change

* ...Delete Dialog Variant Selection By Props

* ...Icon Dynamic, Default Variant Changed

* ...Review Comments Addressed

* ...Run Format Command Changes

* ...Review Comments Addressed

* Add delete student functionality and update modals

Implemented delete student functionality in StudentsList.tsx by introducing a 'DeleteConfirmationModal' component. Renamed 'ConfirmationDialog' to 'DeleteConfirmationModal' and revised it to be more specific, only handling delete confirmation for student records. It's also used in the Admission dashboard. This helps to achieve more precise component function and better code readability.

* Improve delete confirmation modals and function

Added a functional delete button in the 'DeleteConfirmationModal' component in StudentsList.tsx. Updated 'DeleteConfirmationModal' props to be compulsory, removing their optional status. This makes the deletion process more explicit and ensures that modal properties are always provided for better component integrity.

* Add autoFocus to delete button in DeleteConfirmationModal

The autoFocus attribute was added to the "Delete" button in the 'DeleteConfirmationModal' component, improving UI interaction. This change enhances the user experience by ensuring the button is automatically in focus when the modal appears, making deletion process more straightforward and efficient.

---------

Co-authored-by: Arun Selva Kumar <b.arunselvakumar@gmail.com>