Updated feedbacks for remove regulation with toaster (#337)
* Remove debuggers and refactor imports in query files

Debug lines have been removed from 'useDeleteRegulationMutationQuery' and 'useCreateRegulationsMutationQuery' in order to enhance code cleanliness. Additionally, import statements have been reshuffled in several query files, including 'admission', 'regulation', and 'share-form' related files to ensure a more coherent and systematic file structure.

* Update the Toast UI component and RegulationListTable

Refactored the position and z-index properties in the Toast UI component for better positioning style. A new table cell has also been added in the 'RegulationListTable' component that displays whether a regulation is active or not.

* Remove unused CSS class from Tailwind styles

The `.collapsing-row` class was removed from the Tailwind styles as it was no longer in use in the Toast UI and RegulationListTable components. This clean-up helps maintain the quality and efficiency of the code by discarding